requirements.txt
openai
python-dotenv
notebook
ipykernel
pandas
openpyxl
